ok fixing

+++ morphac/constants/src/environment_constants.cc
@@ -3,8 +3,8 @@
 namespace morphac {
 namespace constants {
 
-const double MapConstants::EMPTY{0.};
-const double MapConstants::OBSTACLE{-1.};
+const int MapConstants::EMPTY{0};
+const int MapConstants::OBSTACLE{-1};
 
 }  // namespace constants
 }  // namespace morphac
